rename pyfunc.math to mathfuncs
